+++ Changelog.md
@@ -7,6 +7,13 @@ This project does _not_ adhere to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html) but contrary to it
 every new version is a new major version.
 
+## 124 - 2023-11-08
+
+### Added
+
+-   `className` property to `FileLink` component
+-   `FileLink` will add ellipsis when text overflows
+
 ## 123 - 2023-11-07
 
 ### Added